fixing the issue with warnings from the notifier, with a simpler impl
